fix: protractor.browserstack configuration files have been improved
