+++ protractor.browserstack.live.conf.js
@@ -1,76 +1,30 @@
-// Protractor configuration file, see link for more information
-// https://github.com/angular/protractor/blob/master/lib/config.ts
 
-const { SpecReporter } = require("jasmine-spec-reporter");
-
-if (!process.env.BROWSERSTACK_USERNAME || !process.env.BROWSERSTACK_ACCESS_KEY) {
-    throw new Error("Please add BROWSERSTACK_USERNAME and BROWSERSTACK_ACCESS_KEY " +
-        "to your system`s environment variables.");
-}
-const buildForLocal = ("Local" + (new Date()).getFullYear() + ((new Date()).getMonth() + 1) +
-    (new Date()).getDate() + (new Date()).getHours() + (new Date()).getMinutes());
+const { config } = require("./protractor.browserstack.base.conf.js");
 
 exports.config = {
-    allScriptsTimeout: 1000 * 60 * 2,
-    getPageTimeout: 1000 * 60 * 2,
-    maxSessions: 2,
-    specs: [
-        "./e2e/**/*.e2e-spec.ts"
-    ],
-    commonCapabilities: {
-        "name": "E2E Test",
-        "project": "super-firebase-angular",
-        "build": "" + (process.env.TRAVIS_BUILD_NUMBER || buildForLocal) + "#",
-        "browserstack.user": process.env.BROWSERSTACK_USERNAME,
-        "browserstack.key": process.env.BROWSERSTACK_ACCESS_KEY,
-        "acceptSslCerts": true,
-        "browserstack.local": false,
-        "browserstack.video": false,
-        "browserstack.networkLogs": true
-    },
-    multiCapabilities: [{
-        "browserName": "Chrome"
-    },{
-        "browserName": "Safari"
-    }, {
-        "browserName": "Firefox"
-    }, {
-        "os" : "Windows",
-        "os_version" : "7",
-        "browserName": "IE",
-        "browser_version" : "10.0",
-    }, {
-        "browserName": "android",
-        "device": "Samsung Galaxy S9 Plus",
-        "realMobile": true
-    }, {
-        "browserName": "android",
-        "device": "Google Pixel 2",
-        "realMobile": true
-    }, /*{
-        // https://www.browserstack.com/question/663
-        "browserName": "iPhone",
-        "device": "iPhone X",
-        "realMobile": true
-    }*/],
-    seleniumAddress: "http://hub-cloud.browserstack.com/wd/hub",
-    baseUrl: "https://supermurat-com.firebaseapp.com/",
-    framework: "jasmine",
-    jasmineNodeOpts: {
-        showColors: true,
-        defaultTimeoutInterval: 120000,
-        print() {
-        }
-    },
-    onPrepare() {
-        require("ts-node").register({
-            project: "e2e/tsconfig.e2e.json"
-        });
-        jasmine.getEnv().addReporter(new SpecReporter({spec: {displayStacktrace: true}}));
-    }
+    ...config,
+    baseUrl: "https://supermurat-com.firebaseapp.com/"
 };
+
+exports.config.multiCapabilities.push({
+    // Testing local site with Safari is kind of hard, so let`s test only live site with Safari for now
+    // https://www.browserstack.com/question/663
+    "browserName": "Safari"
+}/*, {
+    // Testing on iPhone is not supported
+    // https://github.com/angular/protractor/issues/2840
+    // https://github.com/angular/protractor/issues/1736
+    // Error: Timeout - Async callback was not invoked within timeout specified by jasmine.DEFAULT_TIMEOUT_INTERVAL.
+    // https://www.browserstack.com/question/663
+    "browserName": "iPhone",
+    "device": "iPhone X",
+    "realMobile": true
+}*/);
 
 // Code to support common capabilities
 exports.config.multiCapabilities.forEach(function(caps){
-    Object.assign(caps, Object.assign({}, exports.config.commonCapabilities, caps));
+    Object.assign(caps, Object.assign({
+            "browserstack.local": false
+        },
+        exports.config.commonCapabilities, caps));
 });
